Fixed stuff
blah blah

+++ Iron Oxide/Assets/Scripts/RepairScripts/Interfaces/IRepairSystem.cs
@@ -4,7 +4,7 @@ public interface IRepairSystem
     public string Partname { get; }
     public bool IsComplete { get; }
     public void Repair();
-    public void StartMinigame(double durability, double durabilityPercentage);
+    public void StartMinigame(float durability, float durabilityPercentage);
     public void StopMinigame();
     public void InterruptMinigame();
 }
